Geometry and Properties are read correct and fast.
